moved ManipDirection/PositionThresh definitions to configurationjitterer.h
